+++ test/Validate/TestObject.cs
@@ -10,6 +10,9 @@ namespace PipServices3.Commons.Test.Validate
         private string PrivateProperty { get; set; } = "XYZ";
         
         public int IntField = 12345;
+        public long LongField = 12345;
+        public float FloatField = 123.45f;
+        public double DoubleField = 123.45;
         public string StringProperty { get; set; } = "ABC";
         public object NullProperty { get; set; } = null;
         public int[] IntArrayProperty { get; set; } = new int[] { 1, 2, 3 };
@@ -17,5 +20,19 @@ namespace PipServices3.Commons.Test.Validate
         public Dictionary<string, int> MapProperty = new Dictionary<string, int>() { { "Key1", 111 }, { "Key2", 222 } };
         public TestSubObject SubObjectProperty { get; set; } = new TestSubObject("1");
         public TestSubObject[] SubArrayProperty { get; set; } = new TestSubObject[] { new TestSubObject("2"), new TestSubObject("3") };
+        public TestEnumInt EnumIntProperty { get; set; } = TestEnumInt.Value0;
+        public TestEnumString EnumStringProperty { get; set; } = TestEnumString.Value1;
+    }
+
+    public enum TestEnumInt
+    {
+        Value0 = 0,
+        Value1 = 1
+    }
+
+    public enum TestEnumString
+    {
+        Value0,
+        Value1
     }
 }